basic agreement

+++ src/main/scala/com/wixpress/academy/raft/package.scala
@@ -2,6 +2,6 @@ package com.wixpress.academy
 
 package object raft {
   type ServerId = Int
-  type EntryIndex = Int
+  type EntryIndex = Long
   type TermIndex = Long
 }
